requirements.txt
fastapi==0.70.1
uvicorn==0.15.0
celery==4.4.7
flower==0.9.7
redis==3.5.3
python-dotenv==0.19.2